:hammer: Refactor booster-module (#62)
* **Refactoring**: The consolidation is focused inside the `config.rs`-file
         - *BoosterIO:* Handles the reading/writing and other IO tasks of the booster.
         - *BoosterConfig:* Handles baseline arguments that are common across `booster.rs` and `multi_output.rs`

* **Renaming**: The `booster.rs` and `multi_output.rs` have been renamed so each file and booster reflects "what it does".
        - `booster.rs` ----> `univariate_booster.rs`
        - `multi_output.rs` ----> `multivariate_booster.rs`

The default-functions and `enums` have been moved to the `config.rs` too. All related calls have been adapted to the new setup. The new interface adds a `cfg`-field to the `UnivariateBooster` and `MultivariateBooster` `structs`. This name was chosen "on the fly", and could be named something else - perhaps something that is more mnemonic.